gbv-sru: switch from TAA to SPR index for language search
TAA only indexes the first language code (i.e. missing results
for items catalogued with several language codes). In addition
it can only be used as an additional filter ANDed to a query
(i.e. search errors in some situations).

The SPR index has the disadvantage that it does not exist on
all indexes for some reason.